fix: makes some copy more consistent and fixes some hot reload issues
