+++ nextjs/src/app/(app)/data-sources/inspect/[externalDataSourceId]/importData.ts
@@ -1,0 +1,48 @@
+import { ApolloClient, FetchResult, gql } from "@apollo/client";
+import { toastPromise } from "@/lib/toast";
+import {
+  ImportDataMutation,
+  ImportDataMutationVariables,
+} from "@/__generated__/graphql";
+
+export default function importData (client: ApolloClient<any>, externalDataSourceId: string) {
+  const importJob = client.mutate<ImportDataMutation, ImportDataMutationVariables>({
+    mutation: gql`
+      mutation ImportData($id: String!) {
+        importAll(externalDataSourceId: $id) {
+          id
+          externalDataSource {
+            importedDataCount
+            isImportScheduled
+            importProgress {
+              status
+              hasForecast
+              id
+              total
+              succeeded
+              failed
+              estimatedFinishTime
+            }
+          }
+        }
+      }
+    `,
+    variables: {
+      id: externalDataSourceId
+    }
+  })
+  toastPromise(importJob, {
+    loading: "Scheduling data import...",
+    success: (d: FetchResult) => {
+      if (!d.errors) {
+        return {
+          title: "Import is processing",
+          description: "This may take a few minutes. You can check the logs for progress."
+        }
+      } else {
+        throw new Error("Couldn't schedule data import")
+      }
+    },
+    error: `Couldn't schedule data import`,
+  });
+}
