+++ nextjs/src/app/reports/[id]/page.tsx
@@ -38,8 +38,8 @@ import { Button, buttonVariants } from "@/components/ui/button";
 import { useRouter } from "next/navigation";
 import spaceCase from 'to-space-case'
 import { toastPromise } from "@/lib/toast";
-import { MAP_REPORT_LAYER_ANALYTICS, ReportMap, selectedConstituencyAtom } from "@/components/report/ReportMap";
-import { MAP_REPORT_FRAGMENT, isConstituencyPanelOpenAtom, isDataConfigOpenAtom } from "@/lib/map";
+import { MAP_REPORT_LAYER_ANALYTICS, ReportMap } from "@/components/report/ReportMap";
+import { MAP_REPORT_FRAGMENT, isConstituencyPanelOpenAtom, isDataConfigOpenAtom, selectedConstituencyAtom } from "@/lib/map";
 import { DisplayOptionsType, ReportContext, defaultDisplayOptions } from "./context";
 import { LoadingIcon } from "@/components/ui/loadingIcon";
 import { contentEditableMutation } from "@/lib/html";
