+++ nextjs/src/components/TopConstituencies.tsx
@@ -5,10 +5,9 @@ import { useContext, useState } from "react"
 import { MemberElectoralInsights, Person } from "./reportsConstituencyItem"
 import { getYear } from "date-fns"
 import { useAtom } from "jotai"
-import { MAX_CONSTITUENCY_ZOOM, selectedConstituencyAtom } from "./report/ReportMap"
+import { MAX_CONSTITUENCY_ZOOM } from "./report/ReportMap"
 import { LoadingIcon } from "./ui/loadingIcon"
-import { useLoadedMap } from "@/lib/map"
-import { constituencyPanelTabAtom } from "@/app/reports/[id]/ConstituenciesPanel"
+import { constituencyPanelTabAtom, selectedConstituencyAtom, useLoadedMap } from "@/lib/map"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "./ui/select"
 import { twMerge } from "tailwind-merge"
